Resolve overlap with main's parallel implementation in favour of this
branch's version (hole detection, QGraphicsView, list widget, mm coords),
while taking main's structural changes (dxflib moved to top-level, double
Node/Model, new files).

Integration fixes:
- Point dxflib include dir at the new top-level dxflib/src.
- Adapt to double Node(x,y,radius)/Model: store hole nodes in mm as
  doubles with radius; AutoWire::WireModel takes double coords.
- inline model_utils::GetDistance (now included by multiple TUs).

Verified: builds (VS2026 + v142), #bulb1.dxf still detects 131 12mm holes.

+++ src/AutoWire.cpp
@@ -6,7 +6,7 @@
 
 #include <algorithm>
 
-void AutoWire::WireModel( int startX, int startY )
+void AutoWire::WireModel( double startX, double startY )
 {
     std::vector< Node > const& nodes = m_model->GetNodes();
